Logging, comments, proxy -> in-memory matcher.
Rewrote to (a) use the SLF4J logging system for debugging logs, and (b) proxy all the matching to the SimpleInMemoryAdaptiveMatcher class.  Also have better error-checking for File objects that don't exist when an object of this class is first created (this shouldnt' necessarily be a problem at the time of the constructor, only if the File can't be saved at close()'s call.)

Also, added some javadoc comments.