Completely reworked logic.
Removed the old buggy implementation of registerMatch(), and replaced with a version that correctly handles multiple sub-, super-, and sibling matches.  Generalization now only occurs if the LUB context is the direct parent of the context of at least one extant or new Match.  Also modified findMatches() to only return parent matches.

Next step: generalize to a "MatchStore."